Tried to rearrange things so that all celery tasks were in a single python file.
