+++ setup.cfg
@@ -90,7 +90,7 @@ banzai_reduce_individual_frame = banzai.main:reduce_single_frame
 banzai_reduce_directory = banzai.main:reduce_directory
 banzai_stack_calibrations = banzai.main:stack_calibrations
 banzai_e2e_stack_calibrations = banzai.main:e2e_stack_calibrations
-banzai_automate_stack_calibrations = banzai.main:start_schedule_calibration_stacking
+banzai_automate_stack_calibrations = banzai.main:start_stacking_scheduler
 banzai_run_realtime_pipeline = banzai.main:run_realtime_pipeline
 banzai_mark_frame_as_good = banzai.main:mark_frame_as_good
 banzai_mark_frame_as_bad = banzai.main:mark_frame_as_bad
